Move mailer service from App folder to Framework folder
